Move tabs-manager to tabs/manager
