+++ app/bg/browser.js
@@ -17,7 +17,7 @@ import { convertDatArchive } from './dat/index'
 import datDns from './dat/dns'
 import { open as openUrl } from './open-url'
 import * as windows from './ui/windows'
-import * as tabManager from './ui/tab-manager'
+import * as tabManager from './ui/tabs/manager'
 import { updateSetupState } from './ui/setup-flow'
 import * as modals from './ui/subwindows/modals'
 import * as siteInfo from './ui/subwindows/site-info'
