+++ app/bg/ui/subwindows/shell-menus.js
@@ -13,7 +13,7 @@ import Events from 'events'
 import { BrowserWindow, BrowserView } from 'electron'
 import * as rpc from 'pauls-electron-rpc'
 import { createShellWindow } from '../windows'
-import * as tabManager from '../tab-manager'
+import * as tabManager from '../tabs/manager'
 import * as modals from './modals'
 import { getToolbarMenu, triggerMenuItemById } from '../window-menu'
 import shellMenusRPCManifest from '../../rpc-manifests/shell-menus'
